Move AssetSuggestions from Assets folder to Ads folder
